change check name; add support for outer and inner multipolygon islands and water body relations
