+++ config/configuration.json
@@ -233,13 +233,12 @@
     },
     "tags":"highway"
   },
-  "WaterbodiesAndIslandSizeCheck": {
-    "enabled": true,
+  "WaterbodyAndIslandSizeCheck": {
     "area":{
       "waterbody.minimum.meters": 10.0,
       "waterbody.maximum.kilometers": 370000.0,
       "island.minimum.meters": 10.0 ,
-      "island.maximum.kilometers": 2130800.0,
+      "island.maximum.kilometers": 2170000.0,
       "islet.maximum.kilometers": 1.0
     }
   }
